Provide MySql5.7Dialect to disable Upsert. Fix Test For ExceptionTest: Class.getDeclaredMethods() The elements in the returned array are not sorted and are not in any particular order.
